Multiple records in a file, PO/Line/Open in one call
