+++ src/main/java/org/olf/folio/order/services/OrderService.java
@@ -17,10 +17,16 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import org.glassfish.jersey.media.multipart.FormDataParam;
+import org.json.JSONArray;
+import org.json.JSONObject;
 import org.olf.folio.order.Constants;
 import org.olf.folio.order.JustMarc;
 import org.olf.folio.order.JustMarcFile;
+import org.olf.folio.order.OrderImport;
+import org.olf.folio.order.OrderImportShortened;
 import org.olf.folio.order.TestImport;
+import org.olf.folio.order.TestImportJustMarc;
+import org.olf.folio.order.TestImportOrderFirst;
 
 
 
@@ -49,11 +55,11 @@ public class OrderService {
 		// SAVE FILE TO DISK
 		writeFile(uploadedInputStream, uploadedFileLocation);
 		// PASS FILE INFO TO 'TestImport' WHICH MAKES THE FOLIO API CALLS
-		TestImportOrderFirst testImport = new TestImportOrderFirst();
+		OrderImportShortened testImport = new OrderImportShortened();
 		testImport.setMyContext(servletRequest.getServletContext());
 		try {
 			//org.json.JSONObject message = testImport.upload(fileDetails.getFileName());
-			org.json.JSONObject message = testImport.upload(fileName.toString() + ".mrc");
+			JSONArray message = testImport.upload(fileName.toString() + ".mrc");
 			return Response.status(Response.Status.OK).entity(message.toString()).build();		
 		}
 		catch(Exception e) {
@@ -74,8 +80,6 @@ public class OrderService {
 	}
 
 
-
-	// save uploaded file to new location
 	private void writeFile(InputStream uploadedInputStream,
 			String uploadedFileLocation) {
 		try {
